% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plcom2012.R
\name{plcom2012}
\alias{plcom2012}
\title{PLCOm2012 risk prediction model for lung cancer}
\usage{
plcom2012(
  age,
  race,
  education,
  bmi,
  copd,
  cancer_hist,
  family_hist_lung_cancer,
  smoking_status,
  smoking_intensity,
  duration_smoking,
  smoking_quit_time
)
}
\arguments{
\item{age}{a vector of patient's age}

\item{race}{categorical variable of patient's race or ethnic group (White, Black, Hispanic,
Asian, American Indian, Alaskan Native, Native Hawaiian, Pacific Islander)}

\item{education}{education was measured in six ordinal levels: less than high-school graduate (level 1),
high-school graduate (level 2), some training after high school (level 3), some college (level 4),
college graduate (level 5), and postgraduate or professional degree (level 6)}

\item{bmi}{a vector of patient's body mass index, per 1 unit of increase}

\item{copd}{binary variable of chronic obstructive pulmonary disease (yes as 1 or no as 0)}

\item{cancer_hist}{binary variable of patient's cancer history (yes as 1 or no as 0)}

\item{family_hist_lung_cancer}{binary variable of patient's family history of lung cancer (yes as 1 or no as 0)}

\item{smoking_status}{binary variable of patient's smoking status (current as 1 or former as 0)}

\item{smoking_intensity}{a vector of the number cigarettes patient smokes per day}

\item{duration_smoking}{a vector of patient's duration of smoking, per 1-yr increase}

\item{smoking_quit_time}{a vector of patient's smoking quit time, per 1-yr increase}
}
\value{
prob patient's 6-year probability of lung-cancer
}
\description{
PLCOm2012 risk prediction model for lung cancer
}
\examples{
plcom2012(age=62, race='White', education=4, bmi=27, copd=0, cancer_hist=0,
family_hist_lung_cancer=0, smoking_status=0, smoking_intensity=80,
duration_smoking=27, smoking_quit_time=10)
}
